+++ crates/apex-core/src/lib.rs
@@ -101,9 +101,9 @@ pub use apex_macros::Component;
 pub use archetype::ArchetypeId;
 pub use commands::Commands;
 pub use component::{
-    make_serde_fns, make_serde_fns_json, Component as ComponentTrait, ComponentHookFn, ComponentId,
-    ComponentRegistry, ComponentSerdeFns, MapEntities, MapEntitiesFn, NoContext, Serializable,
-    SerdeContext, Tick,
+    json_merge, make_serde_fns, make_serde_fns_json, Component as ComponentTrait, ComponentHookFn,
+    ComponentId, ComponentRegistry, ComponentSerdeFns, MapEntities, MapEntitiesFn, NoContext,
+    Serializable, SerdeContext, Tick,
 };
 pub use entity::Entity;
 pub use error::{Anomaly, AnomalyCounts, ErrorHandler, ErrorMode, Severity};